// src/lib/ds/types.ts
export type StringKeyof<T> = Extract<keyof T, string>;
// ...
export type SchemaMemberValue<T> = { [P in StringKeyof<T>]?: T[P] };
export type SchemaMember<T, V> = { [P in StringKeyof<T>]?: V };

export type NewRow<T extends object> = Partial<T> & {
  _ca?: StringKeyof<T>;
  _cid?: string;
  _id?: string;
  _orig?: Partial<T>;
  _newKeys?: StringKeyof<T>[];
  _ov?: unknown;
  _status?: 'N' | 'I';
  _chunkIndex?: number;
  _$select?: string[];
};

export type DBRow<T extends object> = T & {
  _ca?: StringKeyof<T>;
  _changedAttributes?: Partial<T>;
  _cid?: string;
  _id?: string;
  _orig?: Partial<T>;
  _newKeys?: StringKeyof<T>[];
  _ov?: unknown;
  _status: 'Q' | 'U' | 'D' | 'V' | 'E';
  _chunkIndex?: number;
  _$select?: string[];
};

export type Row<T extends object> = NewRow<T> | DBRow<T>;

export type QueryResult<T extends object> = {
  rows: DBRow<T>[];
  fields?: any[];
  elapsed: number;
  sql?: string;
  params?: any[];
  count?: number;
};

export type Query<T extends object> = {
  match?: SchemaMemberValue<T>;
  limit?: number;
  offset?: number;
  select?: StringKeyof<T>[];
  sort?: SchemaMember<T, number>;
  countOnly?: boolean;
  whereClause?: string;
  whereClauseParamList?: any[];
  orderBy?: string;
};